% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_dglm.R
\name{est_dglm}
\alias{est_dglm}
\title{an example for dglm}
\usage{
est_dglm(gamma, beta, n)
}
\arguments{
\item{gamma}{gamma}

\item{beta}{beta}

\item{n}{n}
}
\value{
two estimations
}
\description{
an example for dglm
}
