% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDW.R
\name{PDW}
\alias{PDW}
\title{Weighted Poisson Deviance}
\usage{
PDW(pred, obs, ex = rep(1, length(obs)))
}
\arguments{
\item{pred}{predictions}

\item{obs}{observations}

\item{ex}{weights}
}
\value{
t
}
\description{
Weighted Poisson Deviance
}
