% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_glm.R
\name{poisson_glm}
\alias{poisson_glm}
\title{Weighted GLM estimation for experiment}
\usage{
poisson_glm(data, dep_var_name, explanatory_vars)
}
\arguments{
\item{data}{A dataframe}

\item{dep_var_name}{response variable}

\item{explanatory_vars}{explanatory variables}
}
\value{
A list with two components:
\describe{
\item{m}{The number of iterations taken to converge.}
\item{b}{The estimated coefficients of the model (intercept and slope).}
}
}
\description{
A short description...
}
