% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_glm.R
\name{weighted_glm}
\alias{weighted_glm}
\title{Weighted GLM estimation for poisson distribution}
\usage{
weighted_glm(x, y, w)
}
\arguments{
\item{x}{A numeric vector of independent variable (predictor).}

\item{y}{A numeric vector of dependent variable (response).}

\item{w}{A numeric vector of weights for each observation.}
}
\value{
A list with two components:
\describe{
\item{m}{The number of iterations taken to converge.}
\item{b}{The estimated coefficients of the model (intercept and slope).}
}
}
\description{
A short description...
}
